Moved extractDomain from RecentActivity into NetworkUtils
