+++ RecentActivity/src/org/sleuthkit/autopsy/recentactivity/Firefox.java
@@ -35,6 +35,7 @@ import java.util.logging.Level;
 import org.openide.util.NbBundle;
 import org.sleuthkit.autopsy.casemodule.services.FileManager;
 import org.sleuthkit.autopsy.coreutils.Logger;
+import org.sleuthkit.autopsy.coreutils.NetworkUtils;
 import org.sleuthkit.autopsy.datamodel.ContentUtils;
 import org.sleuthkit.autopsy.ingest.IngestJobContext;
 import org.sleuthkit.autopsy.ingest.IngestServices;
@@ -669,6 +670,6 @@ class Firefox extends Extract {
             return null;
         }
         
-        return Util.extractDomain(url);
+        return NetworkUtils.extractDomain(url);
     }
 }
